requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
yarl==1.9.4
